Created Health Script and Health UI
I've added a player health script and two UI bars, one for health and one for stamina. This can be changed to hearts later if we decide and I'm not sure if we're using stamina but I liked how the bar looked, it can be removed though. I've used the test scene to make these but made them prefabs so if the test scene is overwritten that's okay.