Amelia movement
Added Amelia movement script
created the player movement script ( that i should have named player)
Created the character script taht contains movement and will contain the attack function, the interact funciton, etc. The idea is to use it for master changes to not move that much the player script.
The player script has the get input function and the overide start and update that depend on character.